requirements.txt
appdirs==1.4.4
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
fake-useragent==1.4.0
feedparser==6.0.11
Flask==3.0.2
frozendict==2.4.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
html5lib==1.1
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
MarkupSafe==2.1.5
multitasking==0.0.11
numpy==1.26.4
pandas==2.0.3
parse==1.20.1
peewee==3.17.1
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyee==11.1.0
pyppeteer==2.0.0
pyquery==2.0.0
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
requests-html==0.10.0
rsa==4.9
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.5
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.18
w3lib==2.1.2
webencodings==0.5.1
websockets==10.4
Werkzeug==3.0.1
yahoo-fin==0.8.9.1
yfinance==0.2.36
zipp==3.17.0
